add killOnClose option.
This option provides a way for `ProcessManager` to kill the webpack
watcher process when a `close` message is received from the Electron
main process.

This is convinient because in most cases, when the Electron application
window is closed, it is desired to stop the watcher as well. This patch
makes it possible by opening a comunication channel between the Electron
main process and the watcher process.

Tip: This becomes even more interesting with
`concurrently --kill-others`.